refactor: rename interview_question_category to interview_question_category_dto
