+++ src/google_sheet/dto/interview_question_dto.py
@@ -1,6 +1,8 @@
 from dataclasses import dataclass
 
-from src.google_sheet.dto.interview_question_category import InterviewQuestionCategory
+from src.google_sheet.dto.interview_question_category_dto import (
+    InterviewQuestionCategory,
+)
 from src.google_sheet.dto.interview_question_timestamp_dto import (
     InterviewQuestionTimestamp,
 )
